Handle array providers from env_resolution at the time when they're resolved
Summary:
This diff moves the setup work for empty arrays from the `New_env.init_env` phase into the env resolver. Previously, we would link the providers for empty arrays (as uninhabited tvars) with the type of the empty array itself at the moment when we created the environment entries, but this approach is incompatible with (or at least, doesn't straightforwardly comport with) the sealed tvar environment, because it involves building connections between tvars before their components have been typechecked. Instead, we now treat empty arrays as their own name_def, and do the work of linking their providers at the point when we typecheck that name_def--at which point we will have ensured that the providers themselves have been checked, so we won't be interacting with unpopulated tvars.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D38895890

fbshipit-source-id: 10b243c368c00f3a16fd0a8a3a3ee38a1f9405f6