[contextual] Pass function hint to `Func_sig.toplevels` to construct return hint in the absence of return annotations
Summary:
Changelog: [internal]

There is a mismatch between the statement.ml and name_def.ml hint passing with regards to return hints. name_def will decompose the function hint to get the return hint in the absence of return annotation, but statement.ml will not. This diff fixes the discrepancy by aligning the statement.ml behavior to the behavior of name_def.

Reviewed By: panagosg7

Differential Revision: D38921361

fbshipit-source-id: 6d4138c89b02e5185d5b09e2efe2cabcfdfc9dec