Separately track `id_loc` and `this_write_loc` in `func_info` of class signature
Summary:
Changelog: [internal]

When I first introduced this and super support in the new-env, I went with the path of the least resistence. I observed that the class signature already stored the location of the identifier for class methods, so I directly choose that as the write location of the this in class methods when there is a this annotation.

Recently, I changed the place we bind function this from `Func_sig.toplevels` to `Func_sig.functiontype` and `Func_sig.methodtype`. This change has an unintended effect: `type_annotation.ml` will also call into these functions, so they will also attempt to bind `this`. However, there are no such entries for them to write to. Under reordered-env, this is a harmless useless work, but under enforced-env, it will trigger enforcement error when we try to write to non-existent entries.

In this diff, I finally rectified the design mistake and track two locations separately. Then in statement.ml when we are visiting regular classes, we will pass in the same id_loc as this_write_loc. In type_annotation.ml, we will pass None instead.

In a future diff, I will also standardize the function this write location to always be the location of the entire function for consistency.

Reviewed By: panagosg7

Differential Revision: D38962849

fbshipit-source-id: 2be41365321c313697112ffecc82f4143f24539c