fix get-def of `import *` from `declare module`
Summary:
before, `get-def` on the `Foo` in `import * as Foo from ...` statement fell back on `Get_def_request.Require`, which uses the module resolver to figure out which file to jump to.

this doesn't work if the source is a `declare module`. we have a TODO about implementing this by looking it up in the master context instead:

https://www.internalfb.com/code/fbsource/[35d75c7757b228093e5ab85a7f45414a798a40b2]/fbcode/flow/src/services/get_def/getDef_js.ml?lines=108-114

however, the type of `Foo` is the source module's exports. we want to jump to the exports' def_loc, which we can do with a standard `Get_def_request.Type` instead of `Get_def_request.Require`.

Changelog: [fix] Fix Go to Definition on certain kinds of import statements

Reviewed By: panagosg7

Differential Revision: D38851982

fbshipit-source-id: 312e68a9edc3219d67290d1c0a379a04b8e227dd