fix def-loc of `import *`
Summary:
namespace imports currently use the loc of `Foo` from the `import * as Foo` statement as the loc of the namespace. instead, it should be based on the loc of the exported namespace so that the def loc is correct.

as a result, errors now point through the import which is much more useful since it shows you what the import resolved to. this change is also a pre-requisite to making get-def jump through `import * as Foo`.

Changelog: [misc] Improve location of error references pointing at imported namespaces

Reviewed By: panagosg7

Differential Revision: D38978756

fbshipit-source-id: 8751bb51c539af1d1306262b5b420f8dd426e9a6