[contextual] Use method call hint decomposition for optional member callee
Summary:
Changelog: [internal]]

In a naive implementation of hint decomposition for call expressions that always uses callee as the base hint, we would produce a lot of spurious method-unbinding errors for calls like `foo.bar(...)`, because we would type check `foo.bar` in isolation.

We have already fixed the problem for member callee like `obj.prop` by using `obj` as the base hint and do a method call hint decomposition to get the type of `obj.prop`. However, we omit the case when the callee is an optional member like `obj?.prop(...)`.

In this diff, the case of optional member callee is added and treated the same way as member callee. We don't need extra fancy optional chaining mechanism, because the planned but currently implemented hint decomposition rule that keeps all successful union branches will take care of this.

Reviewed By: mvitousek

Differential Revision: D38912415

fbshipit-source-id: 60bd1187b9212cb928d058fbfda17a95181b860b