[env_builder] Creating AssigningWrites at declaration time instead of hoisting time
Summary:
Changelog: [internal]

We used to create AssigningWrites during hoisting time. There is nothing wrong with this approach with new-env or reordered-env only, since AssigningWrites without any writes to it are harmless. However, they will trigger enforcement errors under LTI.

During hoisting, we know nothing about whether the hoisted binding: it might be illegal or unreachable. Once the AssigningWrite is added, it's impossible to remove during the dead code marker pass.

Therefore, in this diff, I switched the creation of such writes to the declaration sites. Most of them are already created by `pattern_identifier`. The only place we need to add is variable declaration without init but with anno. For unreachable writes, we won't create write entries for them in the main pass, and a NonAssigningWrite will be correctly added in the dead code marker pass, which is what we want.

Reviewed By: mvitousek

Differential Revision: D39002245

fbshipit-source-id: 9c8b4869218506e8324c9982342a341a2980a8e2