Pin this type to `any` when function misses required this annotation under LTI mode
Summary:
Changelog: [internal]

Similar to unannotated parameters under LTI mode, we should pin this type to any when required annotation is missing under LTI mode.

To thread the type to the callsite where we emit the missing annot error, I replaced the bool param `needs_this_param` with a option param `required_this_param_type` that will contain the default this tvar when the annotation is required but missing.

Reviewed By: mvitousek

Differential Revision: D38876194

fbshipit-source-id: 2e3275bcf5cce9cd89665403817684abe0728437