[cleanup] Move `valid_declaration_check` to `name_resolver`
Summary:
Changelog: [internal]

All the info necessary to perform `valid_declaration_check` is already available in the name resolver. The only reason why it is not moved before is to share code with the old env. Now the old env is dead, let's move the code and kill a bunch of now-useless binds!

Reviewed By: gkz

Differential Revision: D39061266

fbshipit-source-id: 1add363a71907662df3c6e1af7a1d7ee4d108ac2