[cleanup] Move `check_predicate_declare_function` to `name_resolver`
Summary:
Changelog: [internal]

All the info necessary to perform `check_predicate_declare_function` is already available in the name resolver. The only reason why it is not moved before is to share code with the old env.

Now the old env is dead, let's move the code. This allows us to kill the entire `statement_decl` phase!

Reviewed By: gkz

Differential Revision: D39061267

fbshipit-source-id: 3b887eace3cae9b2bd425baaaa1ef364b1783a8a