[cleanup] Do not use the reason of the default as init reason during variable assignments
Summary:
Changelog: [internal]

Currently, the use_op produced in destructuring.ml is only used in two places: destructuring assignments and destructuring in variable declaration. These use_ops are used for subtyping against providers, which needs an initializer to begin with, so it make much more sense to use the reason of initializers rather than the default.

Reviewed By: panagosg7

Differential Revision: D39084947

fbshipit-source-id: 54881ad832e4a77a84e10d6fd3fc770f91194f51