Construct positive and negative refinement maps for predicate returns using env builder
Summary:
Changelog: [internal]

In this diff, I implemented the support for creating positive and negative refinement maps for predicate function returns. Previously, this is still constructed by old-env style refinements in statement.ml. This makes us unable to delete more old-env code, and in addition, potentially unsafe since we are not properly invalidating refinements.

In this diff, we will keep track of whether we are in predicate function scope, and if we are in, the parameters of predicate functions. Then at the return, we will synthesize reads after calling `expression_refinement`, thus collecting positive refinements. Then we negate the refinements and then collect the negative refinements.
In env.ml, we will translate the read by converting the refinement read into type predicates, which we already have function to do that. This gives us exactly what we need to construct an `OpenPredT` for predicate function return!

After this diff, more old env code can be killed.

Reviewed By: panagosg7

Differential Revision: D39086331

fbshipit-source-id: 4386fc6682809e674a70f1b4b4d0662f8decc5af