Use global transaction version when compacting shared heap
Summary:
When compacting the heap, we move values based on the global next_version value.
We should also thread entity fields based on the same value.

Later in this stack, I am going to remove all uses of `gc_version_threshold`,
but this use is just unnecessary, so I'm removing it first.

Differential Revision: D38397187

fbshipit-source-id: d5da89922e013ec9829228c918e01951db54c5f1