don't require request ID in LSP tests
Summary:
in the LSP tests, we assert that messages match. but since requests are numbered sequentially and some things happen in parallel, the request IDs aren't always stable.

so this diff ignores the request ID when comparing messages, if the expected message doesn't assert the ID.

we still assert that the messages happen in the right order, though.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D39089755

fbshipit-source-id: 6d5a174af7c87b6ca8a9157c1a6a18fc6ceac0aa