query mergebase directly
Summary:
when an hg.update watchman event finishes, we currently do a watchman query to fetch the new mergebase and the files changed since that new mergebase, but then ignore the changed files. watchman doesn't have a way to query for just the mergebase without the changed files, but we can query source control directly.

Note: when the mergebase changes, we estimate whether it'd be faster to restart than recheck. we estimate how long a restart would take based on how many files are currently checked, which is actually not a good estimate of how many files a restart would have to check. the files a restart would have to check is *precisely* the files changed since the new mergebase! so we should not really be ignoring the files returned here... but since we want to get rid of restarting to save time anyway, I'd rather do this instead.

Changelog: [internal]

Reviewed By: samwgoldman

Differential Revision: D38755078

fbshipit-source-id: 93637b280ff5c0378f64844314cc83213228a8fb