Connect function parameters with function signature under LTI
Summary:
Changelog: [internal]

Before reordered-env, there is a well-established connection between function parameters' bindings and the function signature. The source of truth exists in function signature. Then to obtain the type of function parameters, we will look at the signature, either directly getting a value (in the case of simple identifier parameter), or we do some destructuring (in the case of object and array patterns). In the end, we bind the type of function parameter here:
https://github.com/facebook/flow/blob/3386a975fbadfd828c016224c7406a187a084b85/src/typing/func_stmt_config.ml#L59-L63

In reordered-env, the source of truth becomes the function parameters, since they are considered as "defs" by env_resolution. However, the link between function parameters and function signature is still not broken yet. We will first have the type of the function parameter, then when computing the function signature, we will call the same bind function, but instead of binding the type, it's actually unifying an potentially empty tvar with a type already exists in the environment.

This link is eventually broken in LTI, as we make the bind a noop:
https://github.com/facebook/flow/blob/3386a975fbadfd828c016224c7406a187a084b85/src/typing/env.ml#L739-L742

The goal is to ensure that every env entry is resolved through env_resolution, so this is the behavior we shouldn't change. However, we still need the link.

In this diff, I implemented the following scheme to reestablish the link.

1. We will create a write entry for each parameter at the parameter argument loc (which includes the optional annotation, but excludes the default).
2. During name_def, we will attach the parameter location to bindings that will resolve its type, which are `Annotation` and `Contextual`
3. When resolving `Annotation` and `Contextual`, we will write to the entry at the parameter location.
4. Env_resolution will consider the binding and the param location entry as a whole. When fully resolving tvars for bindings, it will also resolve the tvars in the param loc entry.
5. In statement.ml, we will read the parameter type stored in the function signature, instead of creating new tvars.

This scheme also reveals a bug that we are not declaring the dependency from classes to function defs when class has a field initialized by a function, which I also fixed.

Reviewed By: panagosg7

Differential Revision: D39084948

fbshipit-source-id: 0e89880fa6b360cfac31177dcc331a837de95e64