[EZ] Simplify visiting `Array.isArray` in name_def
Summary:
Changelog: [internal]

Consider the case when we are not visiting `Array.isArray` under non-conditional-context. That's when `cond = NonConditionalContext`. Therefore, we can directly pass cond to the visitor of the argument of `Array.isArray`.

In addition, it's not that we can give any useful hint when we are visiting it in non-conditional context, so it's OK that we are passing Hint_None instead of producing some hint only to be found useless later.

Reviewed By: panagosg7

Differential Revision: D39143646

fbshipit-source-id: ac8c05ffe0898c834d014695d99923dd70596c8a