implement InsertReplaceEdit in completion results
Summary:
Changelog: [feat] When autocompleting in the middle of a word, autocomplete now inserts the suggestion, instead of replacing the entire word. For example, if you are completing in `foo|Baz` (cursor at `|`) and choose `fooBar`, you'll now get `fooBarBaz` instead of `fooBar`. In VS Code, you can press shift-enter to replace instead of insert.

### Background

originally, completions just included the text to insert and the editor had to figure out which range to actually replace. so if you had typed `con|` and accepted `console`, it would insert `sole` and you'd have `console` rather than `conconsole`.

this was fragile and basic, relying on the editor's word splitting, while the language server has a parser. so the LSP added `textEdit`, where the server can provide the range to replace.

flow has always (as far back as I could trace in 2014) used the full token range as the completion range. we only supported autocomplete when the cursor was at the end of a word, so the range before the cursor and the full token range were the same.

in D21180456 (https://github.com/facebook/flow/commit/1c496ecaef87f65641fc17e9c3000b6f85675458) in May 2020, we added support for cursors in the middle of a word. we continued to use the full token range as the completion range, rather than just the prefix range. so if you had something like `element.get|Attribute` and chose `getAnimations`, you'd get `element.getAnimations` not `element.getAnimationsAttribute`. but if you had `myVar`, added `con` to the beginning (`con|myVar`), and tried to complete `console` it'd delete `myVar` instead of helping you write `console.log(myVar)`.

what we do is known as "replace", while the `getAnimationsAttribute` behavior is known as "insert". most other language servers "insert". VS Code's `editor.suggest.insertMode` defaults to "insert".

LSP 3.16 added `InsertReplaceEdit`, a variant of `TextEdit` that lets you provide both insert and replace ranges, so the user can choose. that's what this diff does. the insert range is the prefix before the cursor, and the replace range is the full token.

so, now we match the behavior of other languages, AND you can still get the "replace" behavior by using shift-enter to accept the suggestion, or you can change `editor.suggest.insertMode` to "replace".

Reviewed By: gkz

Differential Revision: D39133511

fbshipit-source-id: ac73c55613747532d4f9b9c43e2f03309fbdd14c