Remove useless binds in `func_stmt_config`
Summary:
Changelog: [internal]

Now that function signature's type is fully connected with function parameters' types, we can remove some useless binds. In addition, instead of doing the same destructuring in destructure.ml that is already done in env_resolution to get the type of parameters, we can directly read the types from the environment in destructuring.ml.

Reviewed By: panagosg7

Differential Revision: D39080177

fbshipit-source-id: 671d603470f8ad3070347f6edee35ae4b1110817