[env_builder] Visit class id after visiting extends
Summary:
Right now, we visit class id before extends in name_resolver. As a result, non-sensical code like `class A extends A {}` type checks. In this diff, I first switched the visiting order in scope_builder. However, this still doesn't fix the issue, because name_resolver still visit id before everything else so that it can do something when the id is None. To support the pattern, I created the `class_identifier_opt` function so that name_resolver can override it. In this way, we can avoid some weird AST rewrites.

The more important motivation of this change is to accurately model the runtime evaluation order, which is important to decide reachability of expressions. Right now, if you do

```
class A extends invariant(false) {}
```

It will still think that `A` is not an unreachable write. I noticed this problem when I'm trying to mark all unreachable writes under LTI mode.

Unfortunately, doing just what's described above will still cause crashes in `class A extends invariant(false) {}`, because statement.ml evaluate things in left to right order. In addition, `class A<T> extends invariant(false) {}` will also crash in env_resolution, because type parameters will always be a write even when it is unreachable. Therefore, we have to create a fallback empty tvar when the class self location entry doesn't exist in the loc env.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D39112078

fbshipit-source-id: 34fa628f466c82d1c3409572045e6ffa14e3a92d