Add assigning write for functions with the same name as declared function
Summary:
Suppose we have a declared function f and a concrete definition of it, we will create a NonAssigningWrite for the concrete definition of f, as the deleted comments says, to avoid a subtype against provider check.

This might be a good scheme back then when we used to have a unify_write_entry function that does both the bind and the subtyping everywhere.

Now that the env.ml module has been mostly cleaned up, it's easy to see that now `init_fun` only does one thing: subtype against providers! Therefore, it's just better to remove `Env.init_fun`, and create AssigningWrite for these functions, so that we don't confuse these functions with unreachable or illegal definitions.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D39124486

fbshipit-source-id: b2e802e99612c40ba84ecd09c376efe08c07d342