Add write barrier to entity_advance
Summary:
Similar to the previous diff, this diff fixes a bug where the incremental GC
might fail to mark a live object.

Entities are a versioned data structure. During a recheck, we do not overwrite
committed data. Instead, we write new data on the side. This makes it possible
to respond to requests using committed data while writing the new data.

When we write new data to an entity, we say we "advance" the entity. This means
we update the entity's version to be >= a global version counter. During the
transaction, such entities have both "latest" and "committed" data. Once the
transaction commits, the "committed" data becomes unreachable.

However, imagine the following scenario:

1. We start an incremental GC, but do not finish marking (phase = mark)
2. We start a recheck (next_version = 2)
3. We advance an entity (entity_version = 3, old = X, new = Y)
4. We read the old pointer X (or a pointner reachable from X) and write that into a new object
5. We commit the recheck (next_version = 4)
6. We resume marking, and now mark the entity from (3)

By the time we mark the entity, the old pointer `X` is unreachable through the
entity, but `X` is still live through the reference in the new object. However,
the new object is not scanned because we allocate black.

Fundamentally, advancing an entity _is_ deleting the old value, except that (1)
it's a logical deletion, we don't actually overwrite the pointer with NULL, and
(2) we don't "delete" it until we commit the transaction.

To address the issue, this diff adds the old value to the mark stack when we
advance the entity, following the intuition that advance = delete.

Reviewed By: mroch

Differential Revision: D39075056

fbshipit-source-id: bf12c676078684dc9147b3b6718e6e65421587fd