Reserve memory during sharedmem init only, not connect as well
Summary:
Pull Request resolved: https://github.com/facebook/flow/pull/8923

We allocate virtual address space, but we don't actually consume memory unless
the space is used.

In Windows, the language is a bit different, as are the constraints. You
"reserve" memory without consuming it, but you must "commit" that memory before
you can access it.

Because we need to commit memory on Windows, we reserve space for the info page,
locals, and hash table. Before this diff, we only did this for Windows, but it's
actually a reasonable thing to do for all platforms.

This diff does two things:
1. Move the code to reserve memory from `define_mappings`, which also runs every
time a worker process connects to shared memory. We only need to do this once,
when we initialize shared memory.
2. Reserve memory unconditionally. We will use posix_fallocate on Linux. In
low-memory situations, this might lead to an early, debuggable OOM exception
instead of a SIGBUS.

In the next diff, I am going to add a new segment for the mark stack. The mark
stack starts out small, but grows as needed. This change makes that code a bit
neater, so I'm taking the opportunity to do this now.

Reviewed By: mroch

Differential Revision: D39075059

fbshipit-source-id: 697017680c4f88501b2bdc71f373b7e9358c9a93