Add write barriers to mutable list cons operations
Summary:
Flow's shared memory GC is incremental, meaning that we do marking work in small
slices, interleaved with rechecks. A recheck can modify the heap, so we need to
be careful to ensure that all live values are marked.

This is a classic problem in incremental garbage collection and it has a classic
solution: write barriers.

For a long time, I had convinced myself that write barriers were unnecessary,
because the data in shared memory was immutable (or, at least, immutable in the
ways that mattered). I was wrong.

For example, we have mutable linked lists:

1. Modules can have multiple provider files, which we represent as an
(intrusive) linked list of files. We add a file to this list by modifying the
head pointer.

2. Modules store their reverse dependencies as a skip list, which is essentially
a sorted linked list. We might insert new list items anywhere in this list.

In both cases, inserting a new list element can cause live objects to be missed
by the marking pass.

For example, imagine a linked list: `X::[]`. Somewhere in memory, we have a head
pointer to the a cons cell containing `X`, which is the end of the list. `X` is
kept alive by the head pointer.

During a recheck, we then allocate a new cons cell, containing `Y` and pointing
to the cons cell containing `X` (that is, we construct `Y::X::[]`), then we
_modify_ the head pointer to point to our new cons cell.

We "allocate black" in shared memory, meaning that all values created during a
marking pass are already considered to be live, so we will not traverse their
fields to find pointers.

So, if we visit the head pointer now, we will visit the cons cell containing Y,
see that it's black, and then stop. We will not visit `X` and it will get
collected.

This diff introduces a write barrier at the moment we modify the head pointer.
Before we update the pointer to `Y::X::[]`, we will add the old pointer (to
`X::[]`) to the mark stack, ensuring that `X` is visited.

Reviewed By: mroch

Differential Revision: D39075057

fbshipit-source-id: 9f906bd747de30c2b56d7b9d082f81db1afae220