Mark unreachable writes related to functions and classes
Summary:
In this diff, I marked functions and classes' related write entries (e.g. class self type and various this and super types) that is not added in the main name_resolver pass as NonAssigningWrite.

However, this is not enough to actually mark them as NonAssigning. In order for the post processing dead code marker to work, AssigningWrite should never be added in the first place during the main pass.

These this and super write entries are considered to be bundled together with the write entry of function and class. If functions are classes are NonAssigning for any reasons, these bundled entries should be considered as non-assigning as well. Thus, I changed the code so that we will only add these extra entries if the functions or classes are assigning writes.

Changelog: [internal]

Reviewed By: mvitousek

Differential Revision: D39117520

fbshipit-source-id: f1a35dec20cd44f4c41a519f9e61b03826d9236a