Only create name-defs for assigning write entries
Summary:
It doesn't make sense to create name-def for non-assigning writes, since those entries will never be read by anything. However, it's not possible to tell from syntax which write entries are assigning in name-def, because name-def is not order-sensitive, and it should definitely not attempt to redo the same analysis in name_resolver.

The best way to ensure that we don't create name-def for non-assigning write entries, is to ensure that we only perform the add, after querying results of name_resolver and found that there was an assigning entry. Previously, this was not possible because a lot of write entries are still being incorrectly marked as AssigningWrite, but now that the issues are fixed after a series of fixes in the previous diffs in the stack, we can turn this on.

Changelog: [internal]

Reviewed By: mvitousek

Differential Revision: D39117519

fbshipit-source-id: fb68dd2d5a21df5012b0646ab167a81d53a65eed