[name_def] Resolving signature of class will also resolve this write entries of function expressions assigned to class properties
Summary:
When we try to evaluate the class signature in `class A { prop = function () {}`, we will call `Func_sig.functiontype`, which will bind the `this` type. This diff recognizes this fact that resolving the class will also resolve these `this` write entries, so that we won't get some FunctionThisLoc not prepared for type checking errors.

Changelog: [internal]

Reviewed By: mvitousek

Differential Revision: D39286983

fbshipit-source-id: 2cf073e714dfa8a57be43702c5388f6b11c69bb9