[name_def] Create name_def for anonymous function declaration
Summary:
It turns out that it's possible to have function declarations without names. Here is an example:

```
export default function () {}
```

This is parsed as `ExportDefault (FunctionDeclaration {id=None, ...})`. Previously, we did not write a write for function declarations without id, because we think this is impossible... This mistake is causing some enforcement errors under LTI because we are not properly resolve the type of `this` write in these functions.

This diff fixes the problem by adding the write for anonymous function declaration.

Changelog: [internal]

Reviewed By: mvitousek

Differential Revision: D39289162

fbshipit-source-id: c52549572aa20b6376f2c512aca0356536190bad