[name_def] Only update function statics state if the write is an AssigningWrite
Summary:
The implementation of function statics has a subtle bug. It assumes that in a statement list, all the function name bindings are unique. This is certainly true for legal programs, but not true for programs with illegal binding.

For example, in the following program:
```
function a() {}
function a() {}
```

Due to the order of the visit, the second definition will override the first one, but name_resolver considers the first one to be legal and the second one illegal. This mismatch is mostly harmless in reordered-env, but under LTI mode, both write entries will be unpopulated.

In this diff, I ensure that we will only update the state if the binding is an AssigningWrite (which means it's legal).

Changelog: [internal]

Reviewed By: gkz

Differential Revision: D39283689

fbshipit-source-id: ece7dd60c6d9687f0697a591f18578ec55143138