[implicit-instantiation] Create a synthesis variant of implicit instantiation solver
Summary:
The current `Pierce` submodule of implicit instantiation solver will error when there are missing bounds. This is not a desirable behavior for contextual typing. Consider the following case:

```
declare function apply<A, B>(A, A=>B): B
apply(3, (n) => String(n))
```

During the initial synthesis, `A` will be pinned to `number`, but `B` will have missing bounds. This will result in an error in `Pierce`, but for the purpose of synthesis, the result is actually good enough: for the targ we care about (`A`), we do have a bound.

Therefore, I created a `Synthesis` variant that won't error but instead put an empty tvar as the inferred type for the solutions missing bounds. This allows contextual typing to go forward unimpeded in most cases, and only results in an empty tvar enforcement error when the solution we do care about is underconstrained.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D39213480

fbshipit-source-id: c4efe5dfd7681f33f433aa0322036f2c5d41ae0e