Error under LTI mode when we failed to evaluate hints
Summary:
In this diff, I enabled missing-local-annot error under LTI when we failed to produce a hint when there is local context available. This happens when hint decomposition fails for whatever reasons. I choose to give it a distinctive error message so that it's easier for us to see how often hint decomposition fails.

Changelog: [internal]

Reviewed By: mvitousek

Differential Revision: D39319540

fbshipit-source-id: 8bbfe3dbb81c9466d0decb4a92d06cf32ffb97ec