[implicit-instantiation] Do not run Pierce under LTI mode when we have context
Summary:
The current Pierce's algorithm running inside flow_js does not have access to the return hint type, which will cause a lot of unactionable unconstrained implicit instantiation errors.

In this diff, I fallback to instantiate all targs with empty tvars under LTI mode when we have return context. In the future, we should either unify the return type hint with the expression when we find the expression type is still underconstrained, or we run Pierce's algorithm in env_resolution where we have access to the return hint.

Changelog: [internal]

Reviewed By: mvitousek

Differential Revision: D39294342

fbshipit-source-id: 1f95afd6a0a49f05868bdab174d043f7659f0f6e