fix watchman reconnection
Summary:
when we lose our connection to watchman, we reconnect. if it's the same instance of watchman, all is well. if watchman restarted, it may have missed some changes and there's some work we need to do to re-sync.

we had this backwards. when watchman restarted, we just reconnected without resyncing. and if the connection just dropped without watchman restarting, we reconnected and resynced, but the resync was a no-op because watchman didn't miss anything.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D39352236

fbshipit-source-id: bda180a90103a9fe47eb0db76460ff57695773c3