[implicit-instantiation] Use bounds of the opposite polarity when we have return hint and we don't find bounds with the same polarity
Summary:
Currently, the implicit instantiation algorithm will complain lack of bounds on `T` in the following simple case when we consider return hints:

```
declare function f(): T
const _: string = f();
```

`T` is in covariant position, so the current algo will use the lower bounds. However, considering the return hint, the only constraints we can establish is `T <: string`, which is an upper bound on `T`.

In this diff, I will make use of lower bounds when we expect upper bounds but upper bounds are not available, and the reverse for upper bounds. This is still sound since targs solver is just a best effort engine to guess what the explicit targs would be, and everything will still be validated during checking anyways. For the purpose of synthesis, having some solutions that might not have the best quality is better than having no solution at all, so this behavior is especially helpful for providing hints on unannotated lambdas in polymorphic function calls.

To limit potential impact on the current post inference workflow, I decide to only enable this behavior when there are return hints, which are currently not passed during the post inference check.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D39213479

fbshipit-source-id: 0c59fb5471f6fefaf4e44dcb4582fd02b45785c2