Allow synthetic names in substitution as long as name couldn't be substituted
Summary:
Originally, `Synthetic` subst names (which occur as the result of representing mutliple generics with the same `GenericT` id, as in the case of spreads) could never be involved in substitution--if we were running `subst` and saw a `GenericT` with a `Synthetic` name, we would crash. This invariant was correct with the classic implicit instantiation mode, because `Synthetic` names only occur as the result of constraints being fired, which always happens after all substitutions could have been made. However, with the switch to Pierce, this is too strong.
```
declare class SearchableEntry<T> {
  constructor(d: T): void
}

function ouch<X>(d: X) {
  new SearchableEntry({ ...d });
}
```
In this above case, when constructing `SearchableEntry`, we determine using Pierce that `T` should be pinned to `{...X}`, the type of `{...d}`, which is represented as a `GenericT` with a `Synthetic` name. We then substitute that generic into `SearchableEntry`. We then, in the standard process of class instantiation, substitute `this` into itself, and this second round of substitution is where we run into trouble--the target of the substitution contains the `Synthetic` `{...X}`.

However, the original point of this invariant was to ensure that we would never need to substitute something that had been "processed" by constraints like spreading; and this remains the case here: the substitution that leads to the crash ("this" => class's tvar) has nothing to do with `{...X}`.

This diff loosens the invariant to still require that (and crash if it's not the case) any name that's being substituted cannot be a component of a `Synthetic` name, which we now know by recording, at the point that `Synthetic` name is created, the non-`Synthetic` names that led to it existing. But if the name being substituted is unrelated to the components of a `Synthertic` name, we're ok.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D39367849

fbshipit-source-id: 6cf08885b20fc64c861126dade4aed83c27cffa9