[implicit-instantiation] Handle ReposLowerT by unwrapping and recursively handle its use_t
Summary:
This diff handles an non-type upper bound ReposLowerT, which will be much more common if we enable the behavior to use the bound from the opposite polarity.

Handling ReposLowerT is easy: it wraps a use_t, so we can unwrap it and recursively call `t_of_use_t` to extract the type.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D39402764

fbshipit-source-id: 65045c1fcddb7aa3e8827a41b375a080ad4e65a5