Include location to annotate in cycle error messages
Summary:
This diff modifies the cycle detection errors to include locations that could be annotated to break the cycle. This information is gathered when detecting cycles--we record the location of function returns and function declarations (by querying the scope builder) to determine where annotations can be inserted.

This isn't trying to do anything clever to detect the minimal set of annotations that need to be inserted--it's simply reporting *all* unannotated positions involved in a cycle.

This enhances the error messages produced, but more importantly it powers the codemod in the follow up diff.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D39414264

fbshipit-source-id: 3fb9d410cff7e5eb3400bd365a69f7e328f49af9