[name_resolver] Fix issue where jsx pragma with member expression causes jsx expressions to be considered unreachable
Summary:
Our dead code marker is not in sync with the main pass of name_resolver when visiting JSX expressions. Currently, it incorrectly assumes that if we are under react runtime classic, then we must have `React.createElement` pragma, so we try to mark the expression is dead is there is no read of React at that location. This messed up all custom jsx pragma member expressions, which will incorrectly record a unreachable read, since in the main pass we correctly decide that there is no read of a name at that location.

In this diff, I fixed the issue by making the dead code marker be in sync with the main pass handling of JSX calls.

Changelog: [errors] JSX expressions in a file with jsx pragma of member expressions (e.g. `Huxpro Foo.bar`) will no longer be incorrectly considered as unreachable. This might uncover previously hidden errors.

Reviewed By: samwgoldman

Differential Revision: D39394244

fbshipit-source-id: 2086a1cdff9a31e8c219761f36b005ccfbfc23b5